pyproject.toml
[project]
name = "page-editor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.27.0",
    "jinja2>=3.1.6",
    "markdown>=3.6",
    "pyyaml>=6.0.2",
    "requests>=2.31.0",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.34.3",
]
